LLM_UI2/requirements.txt
requests
beautifulsoup4
pandas
tqdm
selenium
trafilatura
torch
sentence-transformers
faiss-cpu
openai
urllib3
numpy
einops
streamlit
portkey-ai
